+++ source/include/ir/irbuilder.h
@@ -190,7 +190,9 @@ namespace fir
 
 		[[nodiscard]] Value* SetUnionVariantData(Value* unn, size_t id, Value* data, const std::string& vname = "");
 		[[nodiscard]] Value* GetUnionVariantData(Value* unn, size_t id, const std::string& vname = "");
+
 		[[nodiscard]] Value* GetUnionVariantID(Value* unn, const std::string& vname = "");
+		[[nodiscard]] Value* SetUnionVariantID(Value* unn, size_t id, const std::string& vname = "");